lawn: rename 9P proxy to FS proxy
This same code will be able to be used for SFTP, so make it generic.
